+++ examples/screen.py
@@ -4,18 +4,13 @@ Demonstration of Console.screen()
 
 from time import sleep
 
-from rich.console import Console
 from rich.align import Align
-from rich.text import Text
+from rich.console import Console
 from rich.panel import Panel
 
 console = Console()
 
 with console.screen(style="bold white on red") as screen:
-    for count in range(5, 0, -1):
-        text = Align.center(
-            Text.from_markup(f"[blink]Don't Panic![/blink]\n{count}", justify="center"),
-            vertical="middle",
-        )
-        screen.update(Panel(text))
-        sleep(1)
+    text = Align.center("[blink]Don't Panic![/blink]", vertical="middle")
+    screen.update(Panel(text))
+    sleep(5)