+++ rich/segment.py
@@ -384,9 +384,34 @@ class Segment(NamedTuple):
 
 
 if __name__ == "__main__":  # pragma: no cover
-    lines = [[Segment("Hello")]]
-    lines = Segment.set_shape(lines, 50, 4, style=Style.parse("on blue"))
-    for line in lines:
-        print(line)
+    from rich.syntax import Syntax
+    from rich.text import Text
+    from rich.console import Console
 
-    print(Style.parse("on blue") + Style.parse("on red"))
+    code = """from rich.console import Console
+console = Console()
+text = Text.from_markup("Hello, [bold magenta]World[/]!")
+console.print(text)"""
+
+    text = Text.from_markup("Hello, [bold magenta]World[/]!")
+
+    console = Console()
+
+    console.rule("rich.Segment")
+    console.print(
+        "A Segment is the last step in the Rich render process before gemerating text with ANSI codes."
+    )
+    console.print("\nConsider the following code:\n")
+    console.print(Syntax(code, "python", line_numbers=True))
+    console.print()
+    console.print(
+        "When you call [b]print()[/b], Rich [i]renders[/i] the object in to the the following:\n"
+    )
+    fragments = list(console.render(text))
+    console.print(fragments)
+    console.print()
+    console.print("The Segments are then processed to produce the following output:\n")
+    console.print(text)
+    console.print(
+        "\nYou will only need to know this if you are implementing your own Rich renderables."
+    )